editor/libeditor/tests/test_bug596506.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=596506
-->
<head>
  <title>Test for Bug 596506</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=596506">Mozilla Bug 596506</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 596506 **/

SimpleTest.waitForExplicitFinish();
SimpleTest.waitForFocus(runTest);

const kIsMac = navigator.platform.indexOf("Mac") == 0;


function runTest() {
  var edit = document.getElementById("edit");
  edit.focus();

  sendString("First");
  synthesizeKey("KEY_Enter", {});
  sendString("Second");
  synthesizeKey("KEY_ArrowUp", {});
  synthesizeKey("KEY_ArrowUp", {});
  if (kIsMac) {
    synthesizeKey("KEY_ArrowRight", {accelKey: true});
  } else {
    synthesizeKey("KEY_End", {});
  }
  sendString("ly");
  is(edit.value, "Firstly\nSecond",
     "Pressing end should position the cursor before the terminating newline");
  SimpleTest.finish();
}

</script>
</pre>

<textarea id="edit"></textarea>

</body>
</html>
